♻️ Rename NoteEditorCore to NoteEditor for simplicity
- Renamed component from NoteEditorCore to NoteEditor
- Updated all imports and references in EditorArea and DetachedNoteWindow
- Updated documentation to reflect the new name
- Cleaner naming without unnecessary 'Core' suffix
